Completely remove Starbound references
The SB stuff was fine for testing and experimenting, but the project is
going in a different direction now

It's fine to keep the real-world crops though, since IIRC those are also
in flora (ref. needed). But obviously the SB-specific ones needed to go.

The ones I'm keeping are now in XML to match the potions too, and as an
added bonus I no longer need to use the external jsonpatch lib

TODO: fill in missing descriptions for crops