Remove multiplier
I removed references to the (unused) multiplier from potion.py a few
commits ago, so this is just cleaning up the XML-reading/parsing code.
Reasoning: I think it would be better to not have an ambiguous mutiplier
on effects (does it affect magnitude? duration? both?) and instead
directly change the stats of a potion upon instantiation.